+++ Desktop/ollamadev-ade/boson.config.php
@@ -1,4 +1,4 @@
 <?php
 
-define('OLLAMADEV_VERSION', '0.8.98');
+define('OLLAMADEV_VERSION', '0.8.99');
 define('OLLAMADEV_ADE', true);